+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added
   - New linters:
     - `actionlint` for GitHub Action workflow files
+    - `npm ci --dry-run` for `package-lock.json` files
 
 ## \[0.6.5\] - 2023-07-18